soc: qcom: spss_utils: Add SPSS Hybrid firmware and MSM8998 support
The MSM8998 SoC has a "SPSS Hybrid" firmware, but not a
"SPSS Dev" firmware.
Also, the SPSS_CONFIG_MODE fuses bits don't match with the
ones from SDM845, so we need to select the firmware type
based on different bits for MSM8998: keep both fuses
configurations and differentiate between MSM8998 and
newer SoC by looking at of_machine name: generally, this
contains the SoC name and we don't care about future ones
not reporting theirs there, because we just need 8998 to
get recognized and, in our case, its related DTs are
reporting "qcom,msm8998" as (the/one of) machine compatible.