arm64: DT: Introduce support for SDM845 and SoMC Tama platform
This is an import of the SDM845 DTs from the Open Devices kernel 4.9 LE.UM.2.3.2.r1.4, as of the
latest HEAD.

This does only include SDM845 v2 and v2.1 support, without
the hdk, cdp, qrd (etc) DTs. It also includes some cleanup
to keep the files count lower, since there were some things
that are really not needed.