drm: msm: sde: Initialize merge_3d_count in sde_hw_catalog
This is a bugfix for a bad mistake from Qualcomm: when a device
does not define the merge_3D properties, the merge_3d_count is
never initialized to zero and the function will exit with an
arbitrary value assigned to this field of struct sde_mdss_cfg,
which memory gets allocated without zeroing.

Initialize the merge_3d_count field to zero to allow booting
without DT tricks like assigning invalid configuration to this
section of the catalog.