clk: qcom: dispcc-sdm845: Do not configure the PLL twice
When we probe the driver, we want to run clk_fabia_pll_configure
on the PLL0 of the display clock controller.
In the case of SDM845v2 and SDM670 though, the configure function
is ran twice: first it configures it for SDM845v1, then it does
that again for SDM845v2 or SDM670.

Move the configure call to the fixup function, so that we just
configure it once and correctly.

Signed-off-by: DennySPB <dennyspb@gmail.com>